pyproject.toml
[project]
name = "acp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-client-protocol>=0.8.1",
    "discord-py>=2.3.0",
    "python-dotenv>=1.2.2",
    "pytz>=2026.1.post1",
    "schedule>=1.2.0",
    "tzdata>=2024",
]
